api/requirements.txt
attrs==22.2.0
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
freezegun==1.2.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
pytest==7.2.1
pytest-flask==1.2.0
pytest-freezegun==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.3
typing_extensions==4.5.0
Werkzeug==2.2.3
